+++ src/com/maddyhome/idea/vim/common/Marks.kt
@@ -33,6 +33,14 @@ interface Mark {
 
   fun isClear(): Boolean
   fun clear()
+
+  object KeySorter : Comparator<Mark> {
+    private const val ORDER = "'abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789\"[]^.<>"
+
+    override fun compare(o1: Mark, o2: Mark): Int {
+      return ORDER.indexOf(o1.key) - ORDER.indexOf(o2.key)
+    }
+  }
 }
 
 data class VimMark(