+++ test/org/jetbrains/plugins/ideavim/action/CopyActionTest.java
@@ -146,6 +146,8 @@ public class CopyActionTest extends VimTestCase {
   }
 
   // VIM-476 |yy| |'clipboard'|
+  // TODO: Review this test
+  // This doesn't use the system clipboard, but the TestClipboardModel
   public void testClipboardUnnamed() {
     assertEquals('\"', VimPlugin.getRegister().getDefaultRegister());
     final ListOption clipboardOption = OptionsManager.INSTANCE.getClipboard();
@@ -157,37 +159,38 @@ public class CopyActionTest extends VimTestCase {
                    "<caret>bar\n" +
                    "baz\n");
     final Register starRegister = VimPlugin.getRegister().getRegister('*');
-    if (starRegister != null) {
-      assertEquals("bar\n", starRegister.getText());
-    }
+    assertNotNull(starRegister);
+    assertEquals("bar\n", starRegister.getText());
   }
 
   // VIM-792 |"*| |yy| |p|
+  // TODO: Review this test
+  // This doesn't use the system clipboard, but the TestClipboardModel
   public void testLineWiseClipboardYankPaste() {
     configureByText("<caret>foo\n");
     typeText(parseKeys("\"*yy", "\"*p"));
     final Register register = VimPlugin.getRegister().getRegister('*');
-    if (register != null) {
-      assertEquals("foo\n", register.getText());
-      myFixture.checkResult("foo\n" +
-                            "<caret>foo\n");
-    }
+    assertNotNull(register);
+    assertEquals("foo\n", register.getText());
+    myFixture.checkResult("foo\n" +
+                          "<caret>foo\n");
   }
 
   // VIM-792 |"*| |CTRL-V| |v_y| |p|
+  // TODO: Review this test
+  // This doesn't use the system clipboard, but the TestClipboardModel
   public void testBlockWiseClipboardYankPaste() {
     configureByText("<caret>foo\n" +
                     "bar\n" +
                     "baz\n");
     typeText(parseKeys("<C-V>j", "\"*y", "\"*p"));
     final Register register = VimPlugin.getRegister().getRegister('*');
-    if (register != null) {
-      assertEquals("f\n" +
-                   "b", register.getText());
-      myFixture.checkResult("ffoo\n" +
-                            "bbar\n" +
-                            "baz\n");
-    }
+    assertNotNull(register);
+    assertEquals("f\n" +
+                 "b", register.getText());
+    myFixture.checkResult("ffoo\n" +
+                          "bbar\n" +
+                          "baz\n");
   }
 
   // VIM-1431