+++ CHANGES.md
@@ -23,6 +23,9 @@ To Be Released
 * [VIM-1284](https://youtrack.jetbrains.com/issue/VIM-1284) Fix mapping of digits
 * Fix handling of counts on both operator and motion, e.g. `3d2w` deletes 6 words, instead of 32
 * Allow mapping of `<C-K>` and `<C-V>`/`<C-Q>`
+* [VIM-1899](https://youtrack.jetbrains.com/issue/VIM-1899) Add argument to `:registers` command
+* [VIM-1835](https://youtrack.jetbrains.com/issue/VIM-1835) Macros record input keystrokes instead of mapped keystrokes
+* [VIM-1900](https://youtrack.jetbrains.com/issue/VIM-1900) Ensure non-printable output for `:registers`, `:marks` and `:jumps` is encoded correctly
 
 0.55, 2020-01-20
 --------------